// frontend/src/shared/lib/api/model/cascadeNodeResponse.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Cellar
 * OpenAPI spec version: 0.1.0
 */
import type { CascadeNodeResponseSamplesItem } from './cascadeNodeResponseSamplesItem';

export interface CascadeNodeResponse {
  entity_type: string;
  table: string;
  display_label: string;
  count: number;
  samples: CascadeNodeResponseSamplesItem[];
  truncated: boolean;
  action: string;
  children?: CascadeNodeResponse[];
}